+++ src/Microsoft.Data.SqlClient/netfx/src/Microsoft/Data/SqlClient/TdsParserStateObject.netfx.cs
@@ -100,6 +100,8 @@ namespace Microsoft.Data.SqlClient
             _lastSuccessfulIOTimer = parser._physicalStateObj._lastSuccessfulIOTimer;
         }
 
+        internal SSPIContextProvider CreateSSPIContextProvider() => new NativeSSPIContextProvider();
+
         ////////////////
         // Properties //
         ////////////////
@@ -110,7 +112,7 @@ namespace Microsoft.Data.SqlClient
                 return _sessionHandle;
             }
         }
-        
+
         internal uint Status
         {
             get
