+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/SSPI/NegotiateSSPIContextProvider.cs
@@ -1,0 +1,28 @@
+﻿#if NET7_0_OR_GREATER
+
+using System;
+using System.Text;
+using System.Net.Security;
+
+#nullable enable
+
+namespace Microsoft.Data.SqlClient
+{
+    internal sealed class NegotiateSSPIContextProvider : SSPIContextProvider
+    {
+        private NegotiateAuthentication? _negotiateAuth = null;
+
+        internal override void GenerateSspiClientContext(byte[] receivedBuff, uint receivedLength, ref byte[] sendBuff, ref uint sendLength, byte[][] _sniSpnBuffer)
+        {
+            _negotiateAuth ??= new(new NegotiateAuthenticationClientOptions { Package = "Negotiate", TargetName = Encoding.Unicode.GetString(_sniSpnBuffer[0]) });
+            sendBuff = _negotiateAuth.GetOutgoingBlob(receivedBuff, out NegotiateAuthenticationStatusCode statusCode)!;
+            SqlClientEventSource.Log.TryTraceEvent("TdsParserStateObjectManaged.GenerateSspiClientContext | Info | Session Id {0}, StatusCode={1}", _physicalStateObj.SessionId, statusCode);
+            if (statusCode is not NegotiateAuthenticationStatusCode.Completed and not NegotiateAuthenticationStatusCode.ContinueNeeded)
+            {
+                throw new InvalidOperationException(SQLMessage.SSPIGenerateError() + Environment.NewLine + statusCode);
+            }
+            sendLength = (uint)(sendBuff != null ? sendBuff.Length : 0);
+        }
+    }
+}
+#endif
